# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "乐动数据库服务API文档",
        "title": "Ledong DB API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:31168",
    "basePath": "/api",
    "paths": {
...
        "/prepaidCard/spend": {
            "get": {
                "description": "查询消费记录，支持按会员编号过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消费记录"
                ],
                "summary": "查询消费记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会员编号，可选",
                        "name": "number",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.SpendPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，0表示成功",
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "model.Spend": {
            "type": "object",
            "properties": {
                "annual_times": {
                    "type": "number"
                },
                "charge": {
                    "description": "消费金额",
                    "type": "number"
                },
                "description": {
                    "description": "注意：保持为float32类型",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "quantities": {
                    "type": "integer"
                },
                "times": {
                    "type": "number"
                }
            }
        },
...
        "service.SpendPage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Spend"
                    }
                },
                "first": {
                    "type": "boolean"
                },
                "last": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer"
                },
                "numberOfElements": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}